+++ frontend/src/camera.css
@@ -11,7 +11,6 @@
 .preview {
   display: flex;
   flex-direction: column;
-
   width: 100%;
 }
 
@@ -22,6 +21,10 @@
 .camera-btn-container {
   display: flex;
   justify-content: space-between;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
 }
 
 .btn {
