+++ frontend/src/components/FilePicker.jsx
@@ -1,0 +1,35 @@
+import { useFilePicker } from 'use-file-picker';
+import React from 'react';
+
+export default function FilePicker() {
+  const [openFileSelector, { filesContent, loading, errors }] = useFilePicker({
+    readAs: 'DataURL',
+    accept: 'image/*',
+    multiple: true,
+    limitFilesConfig: { max: 1 },
+    // minFileSize: 0.1, // in megabytes
+    maxFileSize: 50,
+  });
+
+  if (loading) {
+    return <div>Loading...</div>;
+  }
+
+  if (errors.length) {
+    return <div>Error...</div>;
+  }
+
+  return (
+    <div>
+      <button onClick={() => openFileSelector()}>Select files </button>
+      <br />
+      {filesContent.map((file, index) => (
+        <div key={index}>
+          <h2>{file.name}</h2>
+          <img alt={file.name} src={file.content}></img>
+          <br />
+        </div>
+      ))}
+    </div>
+  );
+}
